[client] ongoing refactor-redux
